bboom/requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
exceptiongroup==1.1.2
iniconfig==2.0.0
packaging==23.1
pytest==7.4.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1

